Fixed orphaning bug.
Sometimes, invalidating a renderable did not propagate changes properly.
This was the case if a renderable's parents had no anchor, but regained
them later, or something like that.

I still do not fully understand why this is happening, but the fix seems
sufficient and should not be causing additional problems.